Major refactoring
The script is not in a "usable" state yet. It's going under heavy refactoring and rewrites.
***It may just brick your system if you use as it is now!***
I'm mainly just publishing it now so that I can keep track of changes.
Once it reaches a usable and semi-stable state I will be updating the README to reflect how
the project is used and what the options are.